Moved Audio Machine to own file
